// src/app/shared/interfaces/address.ts
export interface Address {
    ID?: Number;
    default?: boolean;
    firstName: string;
    lastName: string;
    email: string;
    phone: string;
    country: string;
    city: string;
    postcode: string;
    address: string;
    UserID?: number; CompanyName?: string;
    countryCode?: string;
}